hubble: Support `--since` requests in combination with follow-mode
[ upstream commit 0b8148b26ec57da5e237f6ac419b0f6e6c47e091 ]

Previously, the observer implementation assumed time range filters on
the request are not compatible with follow-mode. This however is no
longer the case, we can now apply the since filter when rewinding the
ring buffer. This means that if the user specifies a `since` timestamp,
we first dump all flows newer than `since` before we enter follow-mode.

Fixes: cilium/hubble#363

Signed-off-by: Sebastian Wicki <sebastian@isovalent.com>
Signed-off-by: Ilya Dmitrichenko <errordeveloper@gmail.com>