install: RBAC permissions for finalizers subresources
[ upstream commit 7257d4ae0ab1ad2febc0386051f44b9948736157 ]

Since Cilium sets ownership references on pods, it needs permission
to delete pods via finalizers and for that purpose it also needs
permissions to set the finalizers on pods.
This change is required for OpenShift, however it's based on the GC
admission controller that was introduced in Kubernetes 1.5
(https://github.com/kubernetes/kubernetes/pull/34829).

Also add explicit permissions for finalizers on all CRs, to ensure
that agent and operator can set finalizers on their own resources.

Signed-off-by: Ilya Dmitrichenko <errordeveloper@gmail.com>